third_party/WebKit/LayoutTests/virtual/threaded/fast/scroll-behavior/smooth-scroll/fixed-background-in-iframe.html

<!DOCTYPE html>
<script src="../../../../../resources/js-test.js"></script>
<style>
#scroller {
    position: absolute;
    border: 0;
    left: 100px;
    top: 200px;
    width: 300px;
    height: 300px;
}
</style>
<iframe id="scroller" srcdoc="
  <style>
    div {
      background: radial-gradient(#aaf, white, #aaf) fixed;
      background-size: 200px 200px;
      height: 1000px;
      width: 1000px;
      will-change: transform;
    }
    body {
      margin: 0;
    }
  </style>
  <div></div>
"></iframe>
<script>
var jsTestIsAsync = true;
var iframeDoc;

description("This test verifies that smooth scrolls run on the main thread " +
            "when there are fixed backgrounds on the page.");

// From ScrollAnimatorCompositorCoordinator::RunState.
var RUNNING_ON_MAIN_THREAD = "RunningOnMainThread";
function isScrollAnimatingOnMainThread(node) {
    return internals.getScrollAnimationState(node) == RUNNING_ON_MAIN_THREAD;
}

onload = function() {
    if (!window.eventSender || !window.internals) {
        debug("This test requires window.eventSender.")
        finishJSTest();
        return;
    }
    internals.settings.setScrollAnimatorEnabled(true);

    // Click in the scrollbar.
    eventSender.mouseMoveTo(390, 300);
    eventSender.mouseDown();
    eventSender.mouseUp();

    requestAnimationFrame(function() {
        iframeDoc = document.querySelector("#scroller").contentDocument;
        shouldBeTrue("isScrollAnimatingOnMainThread(iframeDoc)");
        finishJSTest();
    });
};

</script>
